+++ pandapower/control/controller/trafo/DiscreteTapControl.py
@@ -44,7 +44,7 @@ class DiscreteTapControl(TrafoController):
         self.tap_pos = net[self.trafotable].at[tid, "tap_pos"]
         if "vm_set_pu" in kwargs:
             self.vm_set_pu = kwargs["vm_set_pu"]
-            self.vm_delta_pu = self.net[self.trafotable].at[tid, "tap_step_percent"] / 100. * .5
+            self.vm_delta_pu = net[self.trafotable].at[tid, "tap_step_percent"] / 100. * .5 + self.tol
         else:
             self.vm_set_pu = None
             self.vm_delta_pu = None
@@ -60,7 +60,7 @@ class DiscreteTapControl(TrafoController):
 
     def initialize_control(self, net):
         if hasattr(self, 'vm_set_pu') and self.vm_set_pu is not None:
-            self.vm_delta_pu = self.net[self.trafotable].at[tid, "tap_step_percent"] / 100. * .5
+            self.vm_delta_pu = net[self.trafotable].at[self.tid, "tap_step_percent"] / 100. * .5
             self.vm_lower_pu = self.vm_set_pu - self.vm_delta_pu
             self.vm_upper_pu = self.vm_set_pu + self.vm_delta_pu