+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Change Log
 =============
 - [ADDED] Efficient PTDF calculation on large grid
 - [ADDED] toolbox function replace_pq_elmtype()
+- [ADDED] Alternative constructor for DiscreteTapControl to use net.trafo.tap_step_percent to determine vm_lower_pu and vm_upper_pu based on vm_set_pu
 
 [2.6.0]- 2021-03-09
 ----------------------